+++ analyzer/parser/structs.c
@@ -9,7 +9,7 @@
 void fill_struct(void *obj, const StructField *fields, size_t n_fields, const char *data, size_t data_len, const char *sep) {
     for (size_t i = 0; i < n_fields; i++) {
         char *next_sep = memmem(data, data_len, sep, strlen(sep));
-        size_t len = next_sep ? next_sep - data : data_len;
+        size_t len = next_sep ? (size_t)(next_sep - data) : data_len;
 
         char *str = safe_malloc(len + 1);
         memcpy(str, data, len);
